Implement QR Code editor
There's still plenty more I would have liked to have done with this project. But this is what I'll be submitting. I may come back to this project later and continue development.
